From Cyrus Rahman: Allow interface reconfiguration when running in privilege separation mode, document privilege separation

+++ src/racoon/isakmp_inf.c
@@ -169,7 +169,8 @@ isakmp_info_recv(iph1, msg0)
 	if (msg->l < sizeof(*isakmp) + sizeof(*gen)) {
 		plog(LLV_ERROR, LOCATION, NULL, 
 			"ignore information because the "
-			"message is way too short - %d byte(s).\n", msg->l);
+			"message is way too short - %d byte(s).\n",
+			(int)msg->l);
 		goto end;
 	}
 
@@ -196,7 +197,8 @@ isakmp_info_recv(iph1, msg0)
 		if (msg->l < sizeof(*isakmp) + ntohs(gen->len) + sizeof(*nd)) {
 			plog(LLV_ERROR, LOCATION, NULL, 
 				"ignore information because the "
-				"message is too short - %d byte(s).\n", msg->l);
+				"message is too short - %d byte(s).\n",
+				(int)msg->l);
 			goto end;
 		}
